Draft DataFile restructering 2
