+++ include/DataFile.h
@@ -37,6 +37,9 @@ class QTextStream;
 namespace lmms
 {
 
+const unsigned int DATAFILE_VERSION = 24;
+
+
 class ProjectVersion;
 
 
@@ -102,11 +105,15 @@ private:
 
 	void cleanMetaNodes( QDomElement de );
 
+	template<class Up>
+	void upgrade(const unsigned int i, const unsigned int start) {
+		if (i >= start) {
+			Up{*this}();
+		}
+	}
+
 	// helper upgrade routines
-	void upgrade_0_2_1_20070501();
-	void upgrade_0_2_1_20070508();
-	void upgrade_0_3_0_rc2();
-	void upgrade_0_3_0();
+	/*
 	void upgrade_0_4_0_20080104();
 	void upgrade_0_4_0_20080118();
 	void upgrade_0_4_0_20080129();
@@ -115,18 +122,19 @@ private:
 	void upgrade_0_4_0_20080622();
 	void upgrade_0_4_0_beta1();
 	void upgrade_0_4_0_rc2();
+	*/
+
+	/*
 	void upgrade_1_0_99();
 	void upgrade_1_1_0();
 	void upgrade_1_1_91();
 	void upgrade_1_2_0_rc3();
-	void upgrade_1_3_0();
+	*/
+	//void upgrade_1_3_0();
 	void upgrade_noHiddenClipNames();
 	void upgrade_automationNodes();
 	void upgrade_extendedNoteRange();
 	void upgrade_defaultTripleOscillatorHQ();
-	void upgrade_mixerRename();
-	void upgrade_bbTcoRename();
-	void upgrade_sampleAndHold();
 
 	// List of all upgrade methods
 	static const std::vector<UpgradeMethod> UPGRADE_METHODS;
