+++ src/core/datafile/DataFileUpgrade.cpp
@@ -20,9 +20,10 @@
  *
  */
 
-#include "DataFileUpgrade.h"
+#include "datafile/DataFileUpgrade.h"
 
 #include <QDomElement>
+#include <QList>
 
 #include <set>
 #include <cassert>
@@ -32,6 +33,25 @@ namespace lmms
 {
 
 
+// wrappers
+QDomNodeList DataFileUpgrade::elementsByTagName( const QString& tagName ) {
+	return m_document.elementsByTagName(tagName);
+}
+QDomElement DataFileUpgrade::createElement(const QString& tagName) {
+	return m_document.createElement(tagName);
+}
+QDomElement DataFileUpgrade::documentElement() const {
+	return m_document.documentElement();
+}
+QDomElement DataFileUpgrade::firstChildElement(const QString& tagName) const {
+	return m_document.firstChildElement(tagName);
+}
+QDomElement DataFileUpgrade::firstChildElement() const {
+	return m_document.firstChildElement();
+}
+
+
+// helpers
 void DataFileUpgrade::renameElements(QDomNodeList& elements, const QString& newTagName)
 {
 	for (int i = 0; i < elements.length(); ++i)
@@ -52,6 +72,21 @@ void DataFileUpgrade::renameAttribute(QDomNodeList& elements, const QString& old
 		// make before break
 		item.setAttribute(newName, item.attribute(oldName));
 		item.removeAttribute(oldName);
+	}
+}
+
+
+void findIds(const QDomElement& elem, QList<jo_id_t>& idList)
+{
+	if(elem.hasAttribute("id"))
+	{
+		idList.append(elem.attribute("id").toInt());
+	}
+	QDomElement child = elem.firstChildElement();
+	while(!child.isNull())
+	{
+		findIds(child, idList);
+		child = child.nextSiblingElement();
 	}
 }
