+++ src/core/datafile/UpgradeTo03.cpp
@@ -1,0 +1,82 @@
+/*
+ * UpgradeTo03.cpp
+ *   Functor for upgrading data files from pre lmms 0.3.0
+ *
+ * This file is part of LMMS - https://lmms.io
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public
+ * License as published by the Free Software Foundation; either
+ * version 2 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public
+ * License along with this program (see COPYING); if not, write to the
+ * Free Software Foundation, Inc., 51 Franklin Street, Fifth Floor,
+ * Boston, MA 02110-1301 USA.
+ *
+ */
+
+#include "datafile/UpgradeTo03.h"
+
+
+namespace lmms
+{
+
+/*
+ * Upgrade to 0.3.0-rc2
+ *
+ * Upgrade to version 0.3.0-rc2 from some version greater than
+ * or equal to 0.2.1-20070508
+ */
+void UpgradeTo0_3_0_RC2::upgrade()
+{
+	QDomNodeList list = elementsByTagName( "arpandchords" );
+	for( int i = 0; !list.item( i ).isNull(); ++i )
+	{
+		QDomElement el = list.item( i ).toElement();
+		if( el.attribute( "arpdir" ).toInt() > 0 )
+		{
+			el.setAttribute( "arpdir",
+				el.attribute( "arpdir" ).toInt() - 1 );
+		}
+	}
+}
+
+/*
+ * Upgrade to 0.3.0
+ *
+ * Upgrade to version 0.3.0 (final) from some version greater than
+ * or equal to 0.3.0-rc2
+ */
+void UpgradeTo0_3_0::upgrade()
+{
+	QDomNodeList list;
+	while( !( list = elementsByTagName(
+				"pluckedstringsynth" ) ).isEmpty() )
+	{
+		QDomElement el = list.item( 0 ).toElement();
+		el.setTagName( "vibedstrings" );
+		el.setAttribute( "active0", 1 );
+	}
+
+	while( !( list = elementsByTagName( "lb303" ) ).isEmpty() )
+	{
+		QDomElement el = list.item( 0 ).toElement();
+		el.setTagName( "lb302" );
+	}
+
+	while( !( list = elementsByTagName( "channelsettings" ) ).
+							isEmpty() )
+	{
+		QDomElement el = list.item( 0 ).toElement();
+		el.setTagName( "instrumenttracksettings" );
+	}
+}
+
+
+} // namespace lmms
